test(api): remove rds import vpc e2e
